🚚 Rename: 예약 패널 관련 파일 reservation/activity-panel로 이동
